Bug 832788. Part 2: AudioEventTimeline::IsValid needs to accept double. r=ehsan

+++ content/media/webaudio/AudioEventTimeline.h
@@ -72,7 +72,7 @@ private:
     double mDuration;
 
   private:
-    static bool IsValid(float value)
+    static bool IsValid(double value)
     {
       return MOZ_DOUBLE_IS_FINITE(value);
     }
